feat: unify zero-copy parsing and perception visualization
  - add a schema-neutral MessageParser for all built-in zero-copy messages
  - share reflective formatting across dump, eproto, efbs, viewer, and webviz
  - preserve exact 64-bit values and expose parser metadata through Python
  - keep CameraFrame and PointCloud rendering on dedicated fast paths
  - add vmsgs protobuf and FlatBuffers perception mappings
  - support FlatBuffers tables, structs, and object vectors in viewer mappings
  - route lane boundaries and road markings to the correct OSG renderers
  - preserve 64-bit tracking and parking-slot IDs throughout the viewer
  - extend CameraFrame formats and web visualization conversion
  - update tests, documentation, packaging, and changelog

+++ viewer/osg/osgobjectarray.h
@@ -53,7 +53,7 @@ struct ObjectData {
   double velocity[3]{0};
   double score{0};
   uint32_t class_id{0};
-  uint32_t track_id{0};
+  uint64_t track_id{0};
   uint32_t color{0};
   std::string label;
 };
